+++ dinero/__init__.py
@@ -1,6 +1,6 @@
 from ._dinero import Dinero
 
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"
 
 __all__ = ["Dinero"]