+++ dinero/types.py
@@ -1,5 +1,6 @@
-from typing import TypedDict
 from decimal import Decimal
+from typing import TypedDict
+
 from typing_extensions import NotRequired
 
 OperationType = str | int | float | Decimal