+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## [Unreleased](https://github.com/wilfredinni/dinero/compare/0.1.5...master)
+## [Unreleased](https://github.com/wilfredinni/dinero/compare/0.1.6...master)
+
+## [0.1.6](https://github.com/wilfredinni/dinero/releases/tag/0.1.6)
 
 - 100% code coverage 🎉
 - Fixed Type Hints for `multiply` and `divide` methods.